Refactor Transform to CRDT and add webworker (#24)
This change is to make it clear that CRDTs are supported and to improve
the experience of the CRDT algorithms by running them async in a
webworker. This will help not to stall the UI or the Storage when using
larger CRDT models.
The commit also includes some bug fixes and general improvements.

+ Change so that there is a single shared underlying CRDT model for all
identical requests coming from the same user. Unique models is still
supported by setting the crdt conf parameter.

+ Add cursorIndex on CRDT request and response and support in the algos.
This is something which helps with paging.

+ Add model length on CRDT response. This is something which helps with
paging.

+ Add Driver.getNodesById1() function We need this to fetch and return
the added nodes.

+ Add and update tests

+ Add orderByStorageTime feature to AlgoRefId

+ Add CRDT worker thread and browserify build step for it.

+ Add RPC.offCall()

* Improve delta handling and formatting. Delta is sent out as JSON
allowing for more flexibility going forward.

* Update universe.json format to crdt instead of transform Also algos:
number[] is now algo: number, since only a single algo per query is
supported now, dropping chained algos since the added complexity is not
worth it. Adding conf parameter for potential fine tailoring of algos.

* Extract locking from Storage.ts to util/Lock.ts

* Fix bug in Storage.ts about chunking delta

* Simplify AlgoSorted and algoRefId to not use full nodes but simpler
objects containing only the needed data.

* Improve various comments

* Rename allowTransform to allowAlgos

* Miscellaneous code formatting

- Removing MAX LENGTH of CRDT models enforcing since the underlying
fetch query it self is limited already (currently at 1 M rows).

- Remove transfomer.extra field, which is replaced by delta JSON
formatting.

- Remove TRANSFORMER_INVALIDATED response code

---------

Co-authored-by: Maicon Filippsen <filippsen@universe.ai>